Add parameterized env var.
Per ongoing work in flexion/vuls-cloudfoundry-buildpack#2, we do not
want to directly store IP address or hostname of proof-of-concept server
for vuls in our test lab, or anywhere directly in the buildpack at this time.

This parameterized vuls_http_server var is passed in at `cf push` time to
ensure during first build or successive deployments with `cf set-env` and
`cf restage` we can beta-test the plugin without storing sensitive data.